Prepare pip module for a splitting into a package structure
Move the pip module to a package to improve code organization and
maintainability. This change allows for better separation of concerns
and prepares for distribution of functionality into multiple modules
in the future.

Update all test imports to reference the new module location to ensure
tests remain functional.
