+++ hermeto/core/package_managers/pip/__init__.py
@@ -1,0 +1,3 @@
+from hermeto.core.package_managers.pip.main import fetch_pip_source
+
+__all__ = ["fetch_pip_source"]
